+++ CraftMagicItems/CraftMagicItems.csproj
@@ -108,6 +108,7 @@
     <Compile Include="Constants\VisualAdjustmentPatches.cs" />
     <Compile Include="Localization\LocalizationHelper.cs" />
     <Compile Include="Patches\Harmony\ActionBarManagerUpdatePatch.cs" />
+    <Compile Include="Patches\Harmony\ActivatableAbilityOnEventDidTriggerRuleAttackWithWeaponResolvePatch.cs" />
     <Compile Include="Patches\Harmony\AddInitiatorAttackRollTriggerCheckConditionsPatch.cs" />
     <Compile Include="Patches\Harmony\AttackStatReplacementOnEventAboutToTriggerPatch.cs" />
     <Compile Include="Patches\Harmony\BattleLogManagerInitializePatch.cs" />
